Time: 1086 ms (10.00%), Space: 90.8 MB (10.00%) - LeetHub

+++ 2233-maximum-product-after-k-increments/2233-maximum-product-after-k-increments.cpp
@@ -19,6 +19,6 @@ public:
             res = (res * pq.top()) % modNum;
             pq.pop();
         }
-        return (int)res;
+        return res;
     }
 };
